[patch] require users to specify what kind of dev reset to do

+++ packages/node/src/prisma/prisma-database.ts
@@ -29,20 +29,30 @@ export async function doesPrismaDiffExist(
 
 export async function resetDevPrismaDatabase(
     schemaFilePath: string,
+    options: {
+        /**
+         * If you already have migrations created, set this to `true`. If you don't, set it to
+         * `false`. If you don't know which one to use, try both, see which one creates a valid
+         * database for you (try querying it with PrismaClient after running this; if it errors,
+         * this didn't create a valid database).
+         */
+        withMigrations: boolean;
+    },
     env: Record<string, string> = {},
 ) {
-    await runPrismaCommand(
-        {
-            command: 'db push --accept-data-loss --skip-generate',
-            hideLogs: true,
-            ignoreExitCode: true,
-        },
-        schemaFilePath,
-        env,
-    );
-    await runPrismaCommand(
-        {command: 'migrate reset --force --skip-generate --skip-seed'},
-        schemaFilePath,
-        env,
-    );
+    if (options.withMigrations) {
+        await runPrismaCommand(
+            {command: 'migrate reset --force --skip-generate --skip-seed'},
+            schemaFilePath,
+            env,
+        );
+    } else {
+        await runPrismaCommand(
+            {
+                command: 'db push --accept-data-loss --skip-generate',
+            },
+            schemaFilePath,
+            env,
+        );
+    }
 }
